Altered video capture and upload Now h264 mp4 from webcam.

+++ monitor_home/get_lan_devices.sh
@@ -4,9 +4,12 @@ function upload_images {
     then
         for file in $*
         do
-            mod_date=$(date -r "$video_picture_dir/$file" +%Y-%m-%d)
-            ./dropbox_uploader.sh upload $video_picture_dir/$file $video_picture_dir/$mod_date/$file
-            rm $video_picture_dir/$file
+            if ! fuser "$video_picture_dir/$file"; then
+                mod_date=$(date -r "$video_picture_dir/$file" +%Y-%m-%d)
+                ./dropbox_uploader.sh mkdir $video_picture_dir/$mod_date/
+                ./dropbox_uploader.sh upload $video_picture_dir/$file $video_picture_dir/$mod_date/$file
+                rm $video_picture_dir/$file
+            fi
         done
     fi
 }
@@ -214,11 +217,11 @@ fi
 # fi
 
 #Upload any video or photo not uploaded and delete file
-files=$(find $video_picture_dir -type f -printf "%f\n" -name "*.jpeg" -mmin +0)
+files=$(find $video_picture_dir -type f -printf "%f\n" -name "*.jpeg")
 upload_images $files
-files=$(find $video_picture_dir -type f -printf "%f\n" -name "*.mp4" -mmin +0)
+files=$(find $video_picture_dir -type f -printf "%f\n" -name "*.mp4")
 upload_images $files
-files=$(find $video_picture_dir -type f -printf "%f\n" -name "*.mpeg" -mmin +0)
+files=$(find $video_picture_dir -type f -printf "%f\n" -name "*.mpeg")
 upload_images $files
 
 mins=$(date +%M)
